+++ src/plone/restapi/serializer/site.py
@@ -52,7 +52,7 @@ class SerializeSiteRootToJson:
 
         result = {
             # '@context': 'http://www.w3.org/ns/hydra/context.jsonld',
-            "@id": batch.canonical_url,
+            "@id": self.context.absolute_url(),
             "id": self.context.id,
             "@type": "Plone Site",
             "title": self.context.Title(),