+++ src/plone/restapi/tests/test_serializer_catalog.py
@@ -1,4 +1,6 @@
 from DateTime import DateTime
+from pkg_resources import get_distribution
+from pkg_resources import parse_version
 from plone.dexterity.utils import createContentInContainer
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.interfaces import ISerializeToJsonSummary
@@ -8,6 +10,11 @@ from Products.CMFCore.utils import getToolByName
 from zope.component import getMultiAdapter
 
 import unittest
+
+
+HAS_PLONE_6 = parse_version(
+    get_distribution("Products.CMFPlone").version
+) >= parse_version("6.0.0a1")
 
 
 class TestCatalogSerializers(unittest.TestCase):
@@ -43,7 +50,18 @@ class TestCatalogSerializers(unittest.TestCase):
             {"@id": "http://nohost", "items": [], "items_total": 0}, results
         )
 
+    @unittest.skipUnless(HAS_PLONE_6, "Since Plone 6 the Plone site is indexed ...")
     def test_lazy_map_serialization(self):
+        # Test serialization of a Products.ZCatalog.Lazy.LazyMap
+        lazy_map = self.catalog()
+        results = getMultiAdapter((lazy_map, self.request), ISerializeToJson)()
+
+        self.assertDictContainsSubset({"@id": "http://nohost"}, results)
+        self.assertDictContainsSubset({"items_total": 3}, results)
+        self.assertEqual(3, len(results["items"]))
+
+    @unittest.skipIf(HAS_PLONE_6, "... before it was not")
+    def test_lazy_map_serialization_plone5(self):
         # Test serialization of a Products.ZCatalog.Lazy.LazyMap
         lazy_map = self.catalog()
         results = getMultiAdapter((lazy_map, self.request), ISerializeToJson)()