+++ src/plone/restapi/tests/test_site_serializer.py
@@ -1,8 +1,14 @@
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.testing import PLONE_RESTAPI_DX_INTEGRATION_TESTING
+from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
+from plone.restapi.testing import RelativeSession
 from zope.component import getMultiAdapter
 from zope.component import queryUtility
+from plone.app.testing import setRoles
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.app.testing import TEST_USER_ID
 
 import json
 import unittest
@@ -91,3 +97,25 @@ class TestSiteSerializer(unittest.TestCase):
             obj["blocks"]["0358abe2-b4f1-463d-a279-a63ea80daf19"]["url"],
             self.portal.doc1.absolute_url(),
         )
+
+
+class TestSiteSerializationFunctional(unittest.TestCase):
+
+    layer = PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+
+        self.api_session = RelativeSession(f"{self.portal_url}/++api++", test=self)
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+    def tearDown(self):
+        self.api_session.close()
+
+    def test_site_root_get_request(self):
+        response = self.api_session.get("")
+
+        self.assertEqual(response.json()["@id"], self.portal.absolute_url())